pyproject.toml
[project]
name = "gralc-rag"
version = "0.1.0"
description = "Graph-Aware Late Chunking for RAG in Biomedical Literature"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=3.0",
    "torch>=2.4",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "biopython>=1.83",
    "datasets>=2.19",
    "openai>=1.30",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "lxml>=5.1",
    "transformers>=4.40",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
